adicionando interface para separar responsabilidades e adicionando metodo para encriptar senha

+++ internal/controller/model/request/user_request.go
@@ -6,5 +6,5 @@ type UserRequest struct {
 	Email    string `json:"email" binding:"required,email"`
 	Password string `json:"password" binding:"required,gte=6,containsany=!@#$%*"`
 	Name     string `json:"name" binding:"required,gte=4,lte=100"`
-	Age      int8   `json:"age" binding:"required,gte=1,lte=150"`
+	Age      int    `json:"age" binding:"required,gte=1,lte=150"`
 }
